remove Onyx.connect from openReportFromDeepLink and move this function to Link.ts
